Make Wormhole.timeoutMs uint32_t instead of uint16_t. Respect Wormhole.timeoutMs and timeout accordingly. Force the minimum timeout to 1000 ms.

+++ src/bootloader/wormhole.h
@@ -22,7 +22,7 @@
     typedef struct {
       uint64_t magicNumber;
       uint8_t enumerationMode;
-      uint16_t timeoutMs;
+      uint32_t timeoutMs;
     } wormhole_t;
 
 // Variables:
